app/requirements.txt
python-telegram-bot==20.3
numpy==1.24.3
scikit-learn==1.2.2
scipy==1.10.1
requests==2.30.0
matplotlib==3.6.2